Move RuntimeTypeConverter out of pubternal namespace
