+++ src/JsonApiDotNetCore.Annotations/Resources/Identifiable.cs
@@ -1,5 +1,4 @@
 using System.ComponentModel.DataAnnotations.Schema;
-using JsonApiDotNetCore.Resources.Internal;
 
 namespace JsonApiDotNetCore.Resources;
