+++ src/JsonApiDotNetCore/Resources/IdentifiableExtensions.cs
@@ -1,5 +1,4 @@
 using System.Reflection;
-using JsonApiDotNetCore.Resources.Internal;
 
 namespace JsonApiDotNetCore.Resources;
